refactor(app): improve consistency with client crate
